nyan/lexer: move flex dependencies to subdirectory
