+++ nyan/lexer/flex.lpp
@@ -20,7 +20,7 @@
 #endif
 
 #include "lexer.h"
-#include "token.h"
+#include "../token.h"
 
 using namespace std::string_literals;
