+++ nyan/parser.cpp
@@ -6,8 +6,7 @@
 #include "compiler.h"
 #include "database.h"
 #include "file.h"
-#include "flex_lexer.h"
-#include "lexer.h"
+#include "lexer/lexer.h"
 #include "member.h"
 #include "object.h"
 #include "token.h"
